Revert back to global js includes
