+++ docs/index.html
@@ -18,23 +18,15 @@
 <body>
 	<div id="app"></div>
 
-	<!-- Module setup -->
-	<script type="importmap">
-		{
-			"imports": {
-				"vue": "//cdn.jsdelivr.net/npm/vue@3/dist/vue.esm-browser.js",
-				"v-scroll-threshold": "//cdn.jsdelivr.net/npm/v-scroll-threshold@2/dist/v-scroll-threshold.esm.js",
-				"vue-sticky-element": "//cdn.jsdelivr.net/npm/vue-sticky-element@1/dist/vue-sticky-element.esm.js"
-			}
-		}
-	</script>
-	<script type="module">
-		import * as Vue from 'vue';
-		window.Vue = Vue;
-	</script>
-
+	<!-- Vue 2.x -->
+	<script src="//cdn.jsdelivr.net/npm/vue@2/dist/vue.min.js"></script>
+	<!-- <script src="//cdn.jsdelivr.net/npm/vue@2/dist/vue.js"></script> -->
+	<script src="//cdn.jsdelivr.net/npm/v-scroll-threshold@2.0.0/dist/v-scroll-threshold.min.js"></script>
+	<script>var thresholdPlugin = VScrollThreshold;</script>
+	<script src="//cdn.jsdelivr.net/npm/vue-sticky-element@1.1.2/dist/vue-sticky-element.min.js"></script>
 	<!-- Docsify v4 -->
 	<script src="//cdn.jsdelivr.net/npm/docsify@4/lib/docsify.min.js"></script>
+
 	<!-- docsify-themeable (latest v0.x.x) -->
 	<script src="//cdn.jsdelivr.net/npm/docsify-themeable@0/dist/js/docsify-themeable.min.js"></script>
 	<script src="//cdn.jsdelivr.net/npm/prismjs@1.x/components/prism-c.min.js"></script>
@@ -43,6 +35,6 @@
 	<!-- Font Awesome -->
 	<script src="//cdn.jsdelivr.net/gh/zignis/docsify-fa/script.min.js"></script>
 
-	<script type="module" src="app.mjs"></script>
+	<script src="app.js"></script>
 </body>
 </html>
